Car Moving Smoothly
rb.AddForce works as initially intended, some further testing will mean that the physics will be semi realistic for the final product. I have removed the other variables as they are not actually doing anything on run time and it was the Vector2.Lerp which was only doing the movement.